+++ src/DataFrames.jl
@@ -163,8 +163,8 @@ else
 end
 
 const METADATA_FIXED = """
-    Metadata: this function preserves table level and column level metadata
-    that has `:note` style.
+    Metadata: this function preserves table-level and column-level metadata
+    that has `:note`-style.
     """
 
 if VERSION >= v"1.9.0-DEV.1163"